refactor(onboarding): address review feedback on academic screen
- Apply state hoisting: split screen into Stateful/Stateless composables
- Extract `OnboardingDropdownField` to shared components to reduce duplication
- Clear validation errors automatically on user interaction
- Improve UI logic separation and readability
